// models/Character.model.ts
export type CharacterType =
  | "a"
  | "b"
  | "c"
  | "d"
  | "e"
  | "f"
  | "g"
  | "h"
  | "i"
  | "j"
  | "k"
  | "l"
  | "m"
  | "n"
  | "o"
  | "p"
  | "q"
  | "r"
  | "s"
  | "t"
  | "u"
  | "v"
  | "w"
  | "x"
  | "y"
  | "z";

export type CharacterStateType = "?" | "O" | "X" | "N";

export interface CharactersType {
  a: CharacterStateType;
  b: CharacterStateType;
  c: CharacterStateType;
  d: CharacterStateType;
  e: CharacterStateType;
  f: CharacterStateType;
  g: CharacterStateType;
  h: CharacterStateType;
  i: CharacterStateType;
  j: CharacterStateType;
  k: CharacterStateType;
  l: CharacterStateType;
  m: CharacterStateType;
  n: CharacterStateType;
  o: CharacterStateType;
  p: CharacterStateType;
  q: CharacterStateType;
  r: CharacterStateType;
  s: CharacterStateType;
  t: CharacterStateType;
  u: CharacterStateType;
  v: CharacterStateType;
  w: CharacterStateType;
  x: CharacterStateType;
  y: CharacterStateType;
  z: CharacterStateType;
}